使用redis 缓存 会话级分页zset 实现向量数据库下的真分页查询 接口开发

+++ internal/processor/resume_test.go
@@ -57,21 +57,21 @@ func TestPrepareQdrantData_FilterBasicInfo(t *testing.T) {
 			Section: sections[0], // BASIC_INFO
 			Vector:  []float64{0.1, 0.2, 0.3},
 			Metadata: map[string]interface{}{
-				"importance_score": float64(1.0),
+				"importance_score": 1.0,
 			},
 		},
 		{
 			Section: sections[1], // EDUCATION
 			Vector:  []float64{0.4, 0.5, 0.6},
 			Metadata: map[string]interface{}{
-				"importance_score": float64(1.0),
+				"importance_score": 1.0,
 			},
 		},
 		{
 			Section: sections[2], // WORK_EXPERIENCE
 			Vector:  []float64{0.7, 0.8, 0.9},
 			Metadata: map[string]interface{}{
-				"importance_score": float64(1.0),
+				"importance_score": 1.0,
 			},
 		},
 	}
